+++ python/instrumentation/openinference-instrumentation-bedrock/src/openinference/instrumentation/bedrock/_wrappers.py
@@ -79,7 +79,7 @@ class _InvokeModelWithResponseStream(_WithTracer):
 
     @override
     def _wrap_sync(self, wrapped: Callable[..., Any]) -> Any:
-        @wrapt.decorator  # type: ignore[misc]
+        @wrapt.decorator  # type: ignore[misc,attr-defined,unused-ignore]
         def _impl(
             wrapped_fn: Callable[..., Any],
             instance: Any,
@@ -92,7 +92,7 @@ class _InvokeModelWithResponseStream(_WithTracer):
 
     @override
     def _wrap_async(self, wrapped: Callable[..., Any]) -> Any:
-        @wrapt.decorator  # type: ignore[misc]
+        @wrapt.decorator  # type: ignore[misc,attr-defined,unused-ignore]
         async def _impl(
             wrapped_fn: Callable[..., Any],
             instance: Any,
@@ -178,7 +178,7 @@ class _ConverseStream(_WithTracer):
 
     @override
     def _wrap_sync(self, wrapped: Callable[..., Any]) -> Any:
-        @wrapt.decorator  # type: ignore[misc]
+        @wrapt.decorator  # type: ignore[misc,attr-defined,unused-ignore]
         def _impl(
             wrapped_fn: Callable[..., Any],
             instance: Any,
@@ -191,7 +191,7 @@ class _ConverseStream(_WithTracer):
 
     @override
     def _wrap_async(self, wrapped: Callable[..., Any]) -> Any:
-        @wrapt.decorator  # type: ignore[misc]
+        @wrapt.decorator  # type: ignore[misc,attr-defined,unused-ignore]
         async def _impl(
             wrapped_fn: Callable[..., Any],
             instance: Any,
@@ -266,7 +266,7 @@ class _InvokeAgentWithResponseStream(_WithTracer):
 
     @override
     def _wrap_sync(self, wrapped: Callable[..., Any]) -> Any:
-        @wrapt.decorator  # type: ignore[misc]
+        @wrapt.decorator  # type: ignore[misc,attr-defined,unused-ignore]
         def _impl(
             wrapped_fn: Callable[..., Any],
             instance: Any,
@@ -279,7 +279,7 @@ class _InvokeAgentWithResponseStream(_WithTracer):
 
     @override
     def _wrap_async(self, wrapped: Callable[..., Any]) -> Any:
-        @wrapt.decorator  # type: ignore[misc]
+        @wrapt.decorator  # type: ignore[misc,attr-defined,unused-ignore]
         async def _impl(
             wrapped_fn: Callable[..., Any],
             instance: Any,
@@ -354,7 +354,7 @@ class _RetrieveAndGenerateStream(_WithTracer):
 
     @override
     def _wrap_sync(self, wrapped: Callable[..., Any]) -> Any:
-        @wrapt.decorator  # type: ignore[misc]
+        @wrapt.decorator  # type: ignore[misc,attr-defined,unused-ignore]
         def _impl(
             wrapped_fn: Callable[..., Any],
             instance: Any,
@@ -367,7 +367,7 @@ class _RetrieveAndGenerateStream(_WithTracer):
 
     @override
     def _wrap_async(self, wrapped: Callable[..., Any]) -> Any:
-        @wrapt.decorator  # type: ignore[misc]
+        @wrapt.decorator  # type: ignore[misc,attr-defined,unused-ignore]
         async def _impl(
             wrapped_fn: Callable[..., Any],
             instance: Any,