+++ js/packages/openinference-instrumentation-mcp/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @arizeai/openinference-instrumentation-mcp
 
+## 0.2.16
+
+### Patch Changes
+
+- Updated dependencies [cfb128c]
+  - @arizeai/openinference-core@2.1.0
+
 ## 0.2.15
 
 ### Patch Changes