+++ python/instrumentation/openinference-instrumentation-llama-index/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [4.4.1](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-llama-index-v4.4.0...python-openinference-instrumentation-llama-index-v4.4.1) (2026-05-14)
+
+
+### Bug Fixes
+
+* bump openinference-instrumentation minimum to &gt;=0.1.50 ([#3084](https://github.com/Arize-ai/openinference/issues/3084)) ([8a96ad7](https://github.com/Arize-ai/openinference/commit/8a96ad776e723dc1de497b28b25fbdc5e0b12355))
+
+## [4.4.0](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-llama-index-v4.3.10...python-openinference-instrumentation-llama-index-v4.4.0) (2026-05-13)
+
+
+### Features
+
+* **llama_index:** Instrument Sparse Embedding & Workflow Events ([#2928](https://github.com/Arize-ai/openinference/issues/2928)) ([4d03efb](https://github.com/Arize-ai/openinference/commit/4d03efb48951eabe24b46c9476cd5e17934f8cdc))
+
 ## [4.3.10](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-llama-index-v4.3.9...python-openinference-instrumentation-llama-index-v4.3.10) (2026-05-10)